+++ NEWS.md
@@ -5,6 +5,9 @@
 * `find_random()` and `find_random_slopes()` now also extract random effects
   names from auxiliary components.
 
+* `find_random()` and `find_random_slopes()` now include random effects from the
+  dispersion component for models from package *glmmTMB*.
+
 * `clean_parameters()` for brms-models now assigns auxiliary parameters to their
   related `Component` and overwrites former assignments to `"conditional"`.